feat: Add "Forgot Password?" functionality to login screen
This commit adds the functionality to the login screen to allow users to reset their password if they have forgotten it. It introduces a new file, "forgot-password.dart", which contains the implementation for the forgot password dialog. The dialog prompts the user to enter their email address and sends a password reset email to that address when the "Send Reset Email" button is pressed.

Refactor the login.dart file to import the new "forgot-password.dart" file and add a "Forgot Password?" text link to the login screen. When the link is tapped, the forgot password dialog is displayed.

This feature enhances the user experience by providing a convenient way for users to recover their passwords without needing to contact support.